+++ attributes/default.rb
@@ -176,3 +176,5 @@ default['openstack']['dashboard']['password_autocomplete'] = 'on'
 default['openstack']['dashboard']['simple_ip_management'] = false
 default['openstack']['dashboard']['neutron']['enable_lb'] = false
 default['openstack']['dashboard']['neutron']['enable_quotas'] = true
+default['openstack']['dashboard']['neutron']['enable_firewall'] = false
+default['openstack']['dashboard']['neutron']['enable_vpn'] = false