+++ spec/server-redhat_spec.rb
@@ -47,7 +47,7 @@ describe 'openstack-dashboard::server' do
       end
 
       it 'has proper modes' do
-        expect(sprintf('%o', file.mode)).to eq('644')
+        expect(sprintf('%o', file.mode)).to eq('640')
       end
 
       it 'has urls set' do