+++ spec/server_spec.rb
@@ -89,7 +89,7 @@ describe 'openstack-dashboard::server' do
       end
 
       it 'has proper modes' do
-        expect(sprintf('%o', file.mode)).to eq('644')
+        expect(sprintf('%o', file.mode)).to eq('640')
       end
 
       context 'template contents' do