+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of the openstack-dashboar
 ## 10.0.0
 * Upgrading to Juno
 * Upgrading berkshelf from 2.0.18 to 3.1.5
+* Allow enable_filewall and enable_vpn to be configured via attributes
 
 ## 9.1
 * python_packages database client attributes have been moved to the -common cookbook