+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file is used to list changes made in each version of the openstack-dashboar
 * Allow enable_filewall and enable_vpn to be configured via attributes
 * Sync conf files with Juno
 * Add optional section support for local_settings template
+* Update local_settings from 0644 to 0640
 
 ## 9.1
 * python_packages database client attributes have been moved to the -common cookbook