+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This file is used to list changes made in each version of the openstack-dashboar
 * Fix python-ibm-db-django package polluting common package attribute
 * Allow some ceitificate options to be configured
 * Add sensitive flag to private key and certificate file resources
+* Add hash algorithm option to local_settings
 
 ## 9.1
 * python_packages database client attributes have been moved to the -common cookbook